Remove citations that we don't use normatively
I can't find any reference to ISO 19101-1, ISO 19105 throughout the GML encoding files, nor would I really expect to. (I can't find them in the CItyGML-3.0CM repository either)

MIME, URI, 19103, 19111 are used quite a bit; ISO 19109 for "AnyFeature" in core.xsd; 

ISO 19107 Spatial schema & others (ISO 19108, 19111, 19109, 19123) are mentioned in Chuck's "initial population" of a data dictionary of ISO-Classes. The data dictionaries are not currently included in the draft document. Are they needed, or are they better 'by reference' to the Conceptual Model? 

ISO 19108 is mentioned in Annex G, but I don't see that as a normative mention leading to a normative reference.

If the Data Dictionary is not part of this GLM encoding standard, the perhaps an informative reference (bibliography or Annex G.1) to at least ISO 19107 would be helpful.

ISO 19123 was only mentioned here, in the ISO-Classes data dictionary, and in "clause_5_conventions" (discussion of <<Type>>) - and yesterday we discussed removing the Clause 5 UML description.

ISO 19156 was only mentioned here & in the Glossary (as source of a couple of terms that are not actually used)

ISO 19775 is X3D, and we actually cite the W3C/COLLADA standard, not the ISO one.

I have left in ISO 19136-1 and 19136-2 although arguable we could just cite OGC GML 3.2 & 3.3 (the latter, not normatively?)