Remove abbreviations that we don't use
In the "terms" section, should we cite OGC's Observations & Measurements, rather than the ISO one? That said, I think we only use these three terms via SensorThings?

We don't actually use "2D" even in the places where the Conceptual Model does (terrain, relief). A lot of these terms are only used in examples. 

"Terrain intersection curve" is never abbreviated in our document, although it is in the Conceptual Model. We only mention WFS in the bibliography, and not by using the abbreviation.